refactor(ios): consolidate device commands under devx worktree ios
