+++ README.md
@@ -89,13 +89,8 @@ For iPhone Simulator Safari testing, use:
 npx devx worktree simulator
 ```
 
-For Capacitor live reload on a physical iPhone (App DEV scheme), use:
-
-```bash
-npx devx worktree ios-live
-```
-
-Same command is available as `npm run run:ios:live`. See [iOS live reload](#ios-live-reload) below.
+For Capacitor on a physical iPhone, use `npx devx worktree ios <local|prod|live>` (see
+[iOS live reload](#ios-live-reload) for the `live` variant).
 
 When using `npx devx worktree ...` commands, ports are derived from the current worktree path and shown by:
 
@@ -148,7 +143,7 @@ Set in `.env` (or export in your shell):
 ```bash
 EDGE_BIND_HOST=0.0.0.0              # required for physical iPhone access; restart stack after changing
 IOS_LAN_HOST=<mac-lan-ip>           # optional if auto-detect works
-IOS_TARGET_ID=<device-id>           # preferred for npm run run:ios:*
+IOS_TARGET_ID=<device-id>           # preferred for npx devx worktree ios *
 IOS_BACKEND_ORIGIN_LOCAL=http://<mac-lan-ip>:<edge-port>   # optional override
 IOS_BACKEND_ORIGIN_PROD=https://<your-production-host>
 ```
@@ -160,11 +155,12 @@ port plus `IOS_LAN_HOST` (or auto-detected LAN IPv4) when `IOS_BACKEND_ORIGIN_LO
 2. Build, sync, and run on a connected device:
 
 ```bash
-npm run run:ios:prod    # production backend (alias: npm run run:ios)
-npm run run:ios:local   # local Docker edge on your Mac (worktree-aware)
+npx devx worktree ios prod    # production backend (alias: npm run run:ios:prod)
+npx devx worktree ios local   # local Docker edge on your Mac (worktree-aware)
 ```
 
-Connect and trust your iPhone first. `npm run run:ios:*` loads `.env` for `IOS_TARGET_ID` /
+`npm run run:ios:*` are thin aliases for the commands above. Connect and trust your iPhone
+first. `npx devx worktree ios` loads `.env` for `IOS_TARGET_ID` /
 `IOS_TARGET_NAME` (prefer ID). Use `npm run list:ios:targets` to discover values, or
 `npx devx worktree info` to see the configured target. First-time code signing may still
 require opening Xcode once.
@@ -183,7 +179,7 @@ For faster UI iteration on a physical iPhone without rebuilding on every change:
 
 ```bash
 npx devx worktree stack up
-npx devx worktree ios-live    # alias: npm run run:ios:live
+npx devx worktree ios live    # alias: npm run run:ios:live
 ```
 
 Prerequisites:
