+++ package.json
@@ -6,7 +6,6 @@
   "scripts": {
     "prestart": "node scripts/generate-runtime-config.mjs && node scripts/generate-metacritic-platform-artifacts.mjs && node scripts/generate-mobygames-platform-artifacts.mjs",
     "ng": "ng",
-    "start": "ng serve --port 8100 --proxy-config proxy.conf.json",
     "prebuild": "node scripts/generate-runtime-config.mjs && node scripts/generate-metacritic-platform-artifacts.mjs && node scripts/generate-mobygames-platform-artifacts.mjs",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
@@ -40,9 +39,9 @@
     "sync:ios:local": "npm run build:ios:local && npx cap sync ios",
     "sync:ios:prod": "npm run build:ios:prod && npx cap sync ios",
     "run:ios": "npm run run:ios:prod",
-    "run:ios:local": "node scripts/run-ios.mjs local",
-    "run:ios:live": "node scripts/run-ios-live.mjs",
-    "run:ios:prod": "node scripts/run-ios.mjs prod",
+    "run:ios:local": "npx devx worktree ios local",
+    "run:ios:live": "npx devx worktree ios live",
+    "run:ios:prod": "npx devx worktree ios prod",
     "list:ios:targets": "npx cap run ios --list",
     "open:ios": "npx cap open ios"
   },
